smartspace, forecast: adjust weather.com forecast provider and fix inverted lat,lon in the forecast smartspace provider

+++ lawnchair/src/ch/deletescape/lawnchair/smartspace/weathercom/models/SunV1CurrentConditionsResponse.java
@@ -132,7 +132,7 @@ public class SunV1CurrentConditionsResponse extends BaseModel {
         public Integer validTimeGmt;
 
         @SerializedName("vis")
-        public Integer vis;
+        public Double vis;
 
         @SerializedName("wc")
         public Integer wc;
